+++ Tests/SwiftTileJSONTests/OSMExampleTests.swift
@@ -26,7 +26,7 @@ struct OSMExampleTests {
     @Test func osmExampleDecodesFieldsCorrectly() {
         let tileJSON = try! JSONDecoder().decode(TileJSON.self, from: TestData.fromFile("osm")!)
         
-        #expect(tileJSON.tileJSONVersion == Version(3, 0, 0))
+        #expect(tileJSON.tileJSON == Version(3, 0, 0))
         #expect(tileJSON.name == "OpenStreetMap")
         #expect(tileJSON.description == "A free editable map of the whole world.")
         #expect(tileJSON.version == Version(1, 0, 0))